refactor the article page
