+++ frontend/.env
@@ -1,3 +1,4 @@
 # VITE_BACKEND_URL=https://api.projectcatalog.click
 VITE_BACKEND_URL=http://localhost:3000
+VITE_S3_LINK=https://project-catalog-storage.s3.us-east-2.amazonaws.com
